backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
pyyaml==6.0.1
marshmallow==3.20.1
flask-migrate==4.0.5
